continuing with the exercice search logic
