+++ package.json
@@ -74,6 +74,8 @@
 	},
 	"dependencies": {
 		"jwt-decode": "^4.0.0",
+		"lucide": "^0.378.0",
+		"lucide-react": "^0.378.0",
 		"navigate": "^0.3.6",
 		"prop-types": "^15.6.1",
 		"react": "^16.8.4",
