+++ src/front/js/component/User/bodyPartCard.jsx
@@ -1,14 +1,16 @@
 import React from "react";
+import { Dumbbell } from "lucide-react";
+import "../../../styles/User-styles/bodyPartCard.css";
 
-const BodyPartCard = ({ item }) => {
+const BodyPartCard = ({ item, setBodyPart, bodyPart }) => {
   return (
-    <div className="card" style="width: 18rem;">
-      <img src="..." className="card-img-top" alt="..." />
+    <div
+      className="card body-part-card-container"
+      onClick={() => setBodyPart(item)}
+    >
+      <Dumbbell size={50} color="#02b532" strokeWidth={1.75} />
       <div className="card-body">
-        <p className="card-text">
-          Some quick example text to build on the card title and make up the
-          bulk of the card's content.
-        </p>
+        <p className="card-text part-name">{item}</p>
       </div>
     </div>
   );
