+++ src/front/styles/User-styles/horizontalScrollbar.css
@@ -1,0 +1,4 @@
+.horizontal-bar {
+  display: flex;
+  overflow-x: auto;
+}
